Revert "Feat (favorite) : add feature to create FavoriteStation"
This reverts commit fe27b858df52beae636632249ca5ebfa46ad9ef2.
