+++ src/main/java/atdd/path/web/WebConfig.java
@@ -1,4 +1,4 @@
-package atdd;
+package atdd.path.web;
 
 import atdd.Constant;
 import atdd.user.jwt.JwtTokenProvider;
@@ -7,8 +7,6 @@ import atdd.user.web.LoginInterceptor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import static atdd.Constant.*;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
@@ -23,9 +21,6 @@ public class WebConfig implements WebMvcConfigurer {
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
         registry.addInterceptor(new LoginInterceptor(jwtTokenProvider, readProperties))
-                .addPathPatterns(USER_BASE_URI + "/me",
-                        FAVORITE_STATION_BASE_URI+"/**",
-                        FAVORITE_PATH_BASE_URI+"/**"
-                        );
+                .addPathPatterns(Constant.USER_BASE_URI + "/me");
     }
 }
