+++ src/main/java/atdd/user/application/dto/CreateFavoriteStationRequestView.java
@@ -1,0 +1,9 @@
+package atdd.user.application.dto;
+
+public class CreateFavoriteStationRequestView {
+    private Long favoriteStationId;
+
+    public CreateFavoriteStationRequestView(Long favoriteStationId) {
+        this.favoriteStationId = favoriteStationId;
+    }
+}
